Remove FFmpegVideoDecodeEngine's dependency on AVStream.
First step of many towards removing DemuxerStream::QueryInterface, AVStreamProvider, and MediaFormat.

BUG=28206
TEST=media_unittests

Review URL: http://codereview.chromium.org/6624062

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77476 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/video/omx_video_decode_engine.h
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2010 The Chromium Authors. All rights reserved.
+// Copyright (c) 2011 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -176,8 +176,8 @@ class OmxVideoDecodeEngine : public VideoDecodeEngine {
   scoped_refptr<VideoFrame> CreateOmxBufferVideoFrame(
       OMX_BUFFERHEADERTYPE* omx_buffer);
 
-  size_t width_;
-  size_t height_;
+  int width_;
+  int height_;
 
   MessageLoop* message_loop_;
